refactor: terminal in the right spot in the hierarchy
