+++ CTFDeck-app/src/app/app.routes.ts
@@ -1,6 +1,6 @@
 import { Routes } from '@angular/router';
 
-import { TerminalComponent } from './ui/terminal/terminal.component';
+import { TerminalComponent } from '../ui/terminal/terminal.component';
 
 export const routes: Routes = [
   { path: 'terminal', component: TerminalComponent },
