+++ CTFDeck-app/src/ui/terminal/terminal.component.html
